data/pdf/game_of_thrones/requirements.txt
huggingface-hub
accelerate
diffusers
einops
omegaconf
protobuf
pytorch-lightning
safetensors
scipy
tokenizers==0.13.3
transformers
langchain
faiss-cpu
sentence-transformers
pypdf
python-dotenv
bitsandbytes
bs4
compel
sentence_transformers
gradio
torch
torchvision
torchaudio
unstructured
fastapi
fastapi_utils
uvicorn[standard]
sse_starlette
streamlit
spacy
pandasql
llama-cpp-python